+++ test/cpp/catch2/suite1.cpp
@@ -61,3 +61,7 @@ TEST_CASE("throws inside CHECK") {
   CHECK(throws() == 5);
   //
 }
+
+TEST_CASE("试试中文", "[test]") {
+  CHECK(true);
+}